metadata.dtd: Drop obsolete pkgname attribute from {cat,pkg}metadata
This is neither present in the XML schema nor mentioned in GLEP 68.

The pkgname attribute was added in 2003 together with the packages
element, "allowing for a combined view of all metadata files":
https://gitweb.gentoo.org/archive/proj/gentoo.git/commit/?id=b1eae649bb6b7ae623b66cb08d7496f042329f10

In commit 3bbce9c90e430387537c6fdc3d97c3647ed9c465, the package
element has already been removed.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>