2025-09-04-openrc-user-services: new news item
As I remarked on the forums [0], the rollout has been pretty smooth, but
unfortunately a few issues with user configurations have shown up after
an extended period in ~arch (and deployment by other distributions). Explain
to users what is new, how to disable the new thing, and the common problem
we've seen with bad ~/.profile files.

[0] https://forums.gentoo.org/viewtopic-p-8869409.html#8869409

Bug: https://bugs.gentoo.org/952108
Bug: https://bugs.gentoo.org/961693
Bug: https://bugs.gentoo.org/962109
Bug: https://bugs.gentoo.org/962214
Signed-off-by: Sam James <sam@gentoo.org>